Refactored splashScreen.tsx and app.tsx to make authentication work better.
